Simplify installation page, give only 1 conda and 1 pip command.
As discussed in gh-102.  The reason to remove conda-forge is
that `-c conda-forge` isn't best-practice, explaining channels
is too much for this page, and SciPy on Windows is problematic
for conda-forge still so for now defaults is the more useful
channel for the average user.

Keeping `--user` for pip but de-emphasizing it visually so there's
one clear recommended default.