DOC: Fix links
Fixed 7 of 8 broken links in issue #232.

The remaining link is a photo credit to a site that no longer exists.
I tried to find an alternate or updated source for the photo, but
could find no references to the site and could not find the photo
in Google images. I left the link, since we didn't break it.
On that page fixed 'credit' to be consistenly lowercase and to be
singular in the case of a single image.

closes #gh-232