Update content of Data Science tab
Follow up to gh-242, alternative to gh-262. The content rewrite and
inclusion of more relevant libraries attempts to make this sound
natural, sketch the breadth of the Python data science offerings,
and keeps some of the tools like DVC and MLFlow that beginning to
intermediate data scientists really need to learn about.

It does shrink the amount of content to a more reasonable size.