Upload a lighter version of the updated report
Uploading a lighter version of the updated report. I was able to reduce its size to 1,9 MB. 

I wasn’t completely satisfied with the quality of the maps in the original report. Very few people would ever notice, but experts probably would.:) Besides adding one more typeface, the new version of the maps is slightly heavier as it is more detailed, thus an increase in the overall size of the document.