Update NumPy_usersurvey_2020_report.pdf (#488)
* Delete NumPy_usersurvey_2020_report.pdf

Deleting the file to replace with the updated version.

* Update pp. 14-15 of the survey results report

Refined the maps on pp. 14-15: 
– removed the names of the countries with the <1% survey participants, 
– changed the font of the highlighted countries to Overpass (https://fonts.google.com/specimen/Overpass).

* Delete NumPy_usersurvey_2020_report.pdf

Deleting due to the large size of the file, will follow up with a lighter version of the same document.

* Upload a lighter version of the updated report

Uploading a lighter version of the updated report. I was able to reduce its size to 1,9 MB. 

I wasn’t completely satisfied with the quality of the maps in the original report. Very few people would ever notice, but experts probably would.:) Besides adding one more typeface, the new version of the maps is slightly heavier as it is more detailed, thus an increase in the overall size of the document.