Update content to reflect a new policy
This PR  reflects a new policy on including educational content to learn.md. To stay consistent with the values of our project, from now on we will only list the content developed by NumPy contributors and vetted by the community.

I also moved info about the NumPy Tutorials to the Beginners and Advanced sections as it is more fitting with the structure of the page. This decision was discussed with @melissawm and @rossbar who were leading the Tutorials project.